% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Middleton_Island_TUPU}
\alias{Middleton_Island_TUPU}
\title{Data to demonstrate and test multivariate logit Tweedie model}
\usage{
data(Middleton_Island_TUPU)
}
\description{
Data sufficient to demonstrate how to use a Tweedie GLM to provide inference
 about proportions e.g. in food habits analysis, where the model output is
 processed to represent a multivariate logit Tweedie model.
}
\details{
Specifically includes Tufted Puffin bill loads sampled at Middleton Island.

\itemize{
  \item \code{Response} Numeric prey biomass in bill load samples
  \item \code{Year} Numeric year
  \item \code{group} factor representing prey species or category
  \item \code{SampleID} factor with a level for every sampling occasion, e.g., for use in row normalization
}
}
\author{
Mayumi Arimitsu
}
\keyword{data}
