% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mvTweedie.R
\name{mvtweedie}
\alias{mvtweedie}
\title{Multivariate Tweedie distribution for predicting diet proportions}
\description{
Using regression methods to analyze diet proportions for a marked point process
}
\details{
Diet samples often measure a count or biomass for different prey categories.
Rather than converting these data to a proportion and fitting modelling
these proportions, we can instead represent diet samples as an outcome
from a thinnned and double-marked point process, where marks
include prey size per encounter, and prey category, and thinning represents
variation in attack and capture rates is is conceptually similar to
detectability/catchability in other point-count sampling analyses.
Analyzing raw prey measurements (rather than proportions) as a regression
model allows a wide range of models (and associated off-the-shelf software)
and predictions can be converted to proportions (with associated standard
errors) after the model is fitted.  If the prey densities follow a
a Poisson point-process, and prey size per encounter follows a gamma
distribution, then the resulting distribution for biomass of each prey
follows a multivariate Tweedie distribution.
}
\references{
Thorson, J. T., Arimitsu, M. L., Levi, T., & Roffler, G. H. (2022).
Diet analysis using generalized linear models derived from foraging
processes using R package mvtweedie. Ecology, 103(5), e3637.
\doi{10.1002/ecy.3637}
}
