% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mvtweedie.R
\name{predict_mvtweedie}
\alias{predict_mvtweedie}
\title{Get predictions from a multivariate logit interpretation of a Tweedie GLM}
\usage{
predict_mvtweedie(
  model,
  exclude_terms = NULL,
  length_out = 50,
  values = NULL,
  ...
)
}
\description{
It returns a tibble with predicted proportions from a Tweedie GAM model.
}
\details{
The created tibble can then be plotted using \code{ggplot2}
}
