% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mvTweedie.R
\name{predict.mvTweedie}
\alias{predict.mvTweedie}
\title{Predict proportions for new data}
\usage{
\method{predict}{mvTweedie}(x, newdata, origdata = x$frame, se.fit = FALSE)
}
\description{
Predict proportions and associated standard errors using a standard S3 object interface
}
\details{
A Tweedie GLM using a log-link and multiple categories can be
transformed to yield predicted proportions and associated SEs, where the
model is interpreted as a multivariate logit Tweedie distribution. This function
does this transformation for a model fitted using:
\itemize{
\item A gam using \code{\link[mgcv]{gam}}
\item A GLMM using \code{\link[glmmTMB]{glmmTMB}}
}
}
