% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mvtweedie.R
\name{predict.mvtweedie}
\alias{predict.mvtweedie}
\title{Predict proportions for new data}
\usage{
\method{predict}{mvtweedie}(
  x,
  category_name = "group",
  newdata,
  origdata = x$frame,
  se.fit = FALSE
)
}
\arguments{
\item{x}{output from \code{\link[mgcv]{gam}} or \code{\link[glmmTMB]{glmmTMB}}, but with
\code{class(x)=c("mvtweedie",...)} where \code{...} indicates the original values for
\code{class(x)}}

\item{category_name}{name of column that indicates grouping variable}

\item{newdata}{An optional data frame in which to look for variables with
    which to predict.  If omitted, the fitted values are used.}

\item{origdata}{original data used when fitting}

\item{se.fit}{A switch indicating if standard errors are required.}
}
\description{
Predict proportions and associated standard errors using a standard S3 object interface
}
\details{
A Tweedie GLM using a log-link and multiple categories can be
transformed to yield predicted proportions and associated SEs, where the
model is interpreted as a multivariate logit Tweedie distribution. This function
does this transformation for a model fitted using:
\itemize{
\item A generalized additive model (GAM) using \code{\link[mgcv]{gam}}
\item A generalized linear mixed model (GLMM) using \code{\link[glmmTMB]{glmmTMB}}
}
It then also calculates an approximation to the standard error for this proportion
}
\examples{
\dontrun{
# Load packages
library(mvtweedie)

# load data set
data( Middleton_Island_TUPU, package="mvtweedie" )
DF = Middleton_Island_TUPU

# Run Tweedie GLM
gam0 = gam( formula = Response ~ 0 + group, data = DF, family = tw )

# Inspect results
class(gam0) = c( "mvtweedie", class(gam0) )
predict(gam0, se.fit=TRUE, origdata = DF)
}


}
